+++ ResultsSet.hpp
@@ -34,7 +34,7 @@ public:
     
     // return normal result or rank 
     // depending on param settings
-    int res(size_t iIdx, size_t iAlg) const;
+    double res(size_t iIdx, size_t iAlg) const;
 
     const std::vector<std::string> &algsettings() const {
         return this->algsettings_;
